Calculator codes

+++ subtraction.py
@@ -1,4 +1,4 @@
 num1 = input('enter first number:')
 num2 = input('enter second number:')
-difference = float(num1) - float(num2)
+difference = int(num1) - int(num2)
 print('The difference between {0} and {1} is {2}'. format(num1, num2, difference))
